requirements.txt
absl-py==2.3.1
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.2.0
ccxt==4.5.40
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
coincurve==21.0.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dlib==19.24.8
docutils==0.21.2
exceptiongroup==1.3.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.116.2
filelock==3.19.1
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.60.0
frozenlist==1.6.0
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.75.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.3
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.4.4
namex==0.1.0
narwhals==2.5.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pillow==10.2.0
propcache==0.3.1
protobuf==6.32.1
pyarrow==21.0.0
pycares==5.0.1
pycparser==3.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.4
pypiwin32==223
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==20.6
python-whois==0.9.5
pytz==2025.2
pywin32==310
referencing==0.36.2
requests==2.32.3
rich==14.1.0
rpds-py==0.27.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.48.0
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.0
